Make Tau names more consistent case-wise
