+++ osu.Game.Rulesets.Tau.Tests/osu.Game.Rulesets.Tau.Tests.csproj
@@ -18,7 +18,7 @@
     <PackageReference Update="Microsoft.EntityFrameworkCore.Sqlite" Version="3.1.4" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\osu.Game.Rulesets.tau\osu.Game.Rulesets.Tau.csproj" />
+    <ProjectReference Include="..\osu.Game.Rulesets.Tau\osu.Game.Rulesets.Tau.csproj" />
   </ItemGroup>
   <PropertyGroup Label="Project">
     <OutputType>WinExe</OutputType>
