+++ osu.Game.Rulesets.Tau/Beatmaps/TauBeatmap.cs
