+++ osu.Game.Rulesets.Tau/Beatmaps/TauBeatmapConverter.cs
