+++ osu.Game.Rulesets.Tau/Configuration/TauRulesetConfigManager.cs
