+++ osu.Game.Rulesets.Tau/Extensions.cs
