+++ osu.Game.Rulesets.Tau/Mods/TauModAutoplay.cs
