+++ osu.Game.Rulesets.Tau/Mods/TauModDoubleTime.cs
