+++ osu.Game.Rulesets.Tau/Mods/TauModEasy.cs
