+++ osu.Game.Rulesets.Tau/Mods/TauModFlashlight.cs
