+++ osu.Game.Rulesets.Tau/Mods/TauModHalfTime.cs
