+++ osu.Game.Rulesets.Tau/Mods/TauModHardRock.cs
