+++ osu.Game.Rulesets.Tau/Objects/TauHitObject.cs
