+++ osu.Game.Rulesets.Tau/Replays/TauAutoGenerator.cs
