+++ osu.Game.Rulesets.Tau/Resources/Textures/tau.png
