+++ osu.Game.Rulesets.Tau/TauRuleset.cs
@@ -29,7 +29,7 @@ namespace osu.Game.Rulesets.Tau
         public override string Description => "tau";
 
         public override DrawableRuleset CreateDrawableRulesetWith(IBeatmap beatmap, IReadOnlyList<Mod> mods = null) =>
-            new DrawabletauRuleset(this, beatmap, mods);
+            new DrawableTauRuleset(this, beatmap, mods);
 
         public override IBeatmapConverter CreateBeatmapConverter(IBeatmap beatmap) =>
             new TauBeatmapConverter(beatmap, this);
