+++ osu.Game.Rulesets.Tau/UI/KiaiHitExplosion.cs
